requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.16.0
PyJWT==2.5.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
whitenoise==6.2.0
